+++ objectbox/lib/src/native/bindings/objectbox.h
@@ -52,7 +52,7 @@ extern "C" {
 /// When using ObjectBox as a dynamic library, you should verify that a compatible version was linked using
 /// obx_version() or obx_version_is_at_least().
 #define OBX_VERSION_MAJOR 4
-#define OBX_VERSION_MINOR 2
+#define OBX_VERSION_MINOR 3
 #define OBX_VERSION_PATCH 0  // values >= 100 are reserved for dev releases leading to the next minor/major increase
 
 //----------------------------------------------
@@ -669,7 +669,12 @@ typedef enum {
     /// JavaScript source code
     /// Representing type: String
     OBXExternalPropertyType_JavaScript = 111,
-    // OBXExternalPropertyType_Reserved5 = 112,
+    /// A JSON string that is converted to a native representation in the external system.
+    /// For example, a JSON object on the ObjectBox side (string) would be converted to an embedded document in MongoDB.
+    /// It depends on the external system what kind of JSON structures is supported.
+    /// For MongoDB, this is very flexible and allows (nested) objects, arrays, primitives, etc.
+    /// Representing type: String
+    OBXExternalPropertyType_JsonToNative = 112,
     // OBXExternalPropertyType_Reserved6 = 113,
     // OBXExternalPropertyType_Reserved7 = 114,
     // OBXExternalPropertyType_Reserved8 = 115,