+++ objectbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## latest
 
+* Update ObjectBox database for Flutter Linux/Windows, Dart Native apps to [4.3.0](https://github.com/objectbox/objectbox-c/releases/tag/v4.3.0).
+* Update ObjectBox database for Flutter Android apps to 4.3.0.
+  If your project is [using Admin](https://docs.objectbox.io/data-browser#admin-for-android), make
+  sure to update to `io.objectbox:objectbox-android-objectbrowser:4.3.0` in `android/app/build.gradle`.
+* Update ObjectBox database for Flutter iOS/macOS apps to 4.3.0.
+  For existing projects, run `pod repo update` and `pod update ObjectBox` in the `ios` or `macos` directories.
+* External property types (via [MongoDB connector](https://sync.objectbox.io/mongodb-sync-connector)):
+  add `jsonToNative` to support sub (embedded/nested) documents/arrays in MongoDB.
+
 ## 4.2.0 (2025-04-15)
 
 * Requires at least Dart SDK 3.4 or Flutter SDK 3.22.